+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Actions das chamadas websocket nos docs [#447](https://github.com/3C-gg/reload-backend/issues/447).
 - Envio de WS de expiração de convites quando usuário fizer logout `invites/expire`.
 - Envio de WS quando usuário se cadastrar `friends/create` [#444](https://github.com/3C-gg/reload-backend/issues/444).
 - Campo `profileurl` nos dados de inicialização pra testes locais (`seed.json`).